Dylan's improvements of weight parsing
changed LHEsource so it compiles, running tests now

added scientific notation for numbers so values are correct

added a few more test statements. Need to think of future needs

Add testing suite (will remove/squash commits right before merge)

Change pdfSetsInfo to be read in from LHAPDF path
Still lingering question of how to decide pdf type
This can be resolved in the new function "setupPdfSetsInfo"

Add check if orphan weight is pdf and fix xml tag swap (needs review)

Allow reading '&gt;' as '>' and fix bug with first weight w/o group
As for the first weight error: if the first weight doesn't have a
weight group, the code would not add it to the list of weights because
of currentGroupName and the first weight both having a trivial "" as
their name. Thats the reason for the currentGroupName being set to
"None" for initialization

Fix space issue in regex and moved unchaning variables outside for

Add pdfgroup splitting. To do this, needed LHAPDF requirements

clang format